refactor: extract a common function create_cgroup_manager
